requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Embedding models
openai>=1.0.0
cohere>=4.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Korean language models
kobert-transformers>=0.5.1

# Vector stores
chromadb>=0.4.0
faiss-cpu>=1.7.4
pinecone>=5.0.0
qdrant-client>=1.7.0

# New vector stores (Docker Compose required)
pymilvus>=2.3.0
weaviate-client>=4.4.0
elasticsearch>=8.11.0
psycopg2-binary>=2.9.9
redis>=5.0.0

# Evaluation metrics
scikit-learn>=1.3.0
nltk>=3.8.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# PDF processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
